new try

+++ treemap.c
@@ -59,7 +59,7 @@ TreeMap * createTreeMap(int (*lower_than) (void* key1, void* key2)) {
 Pair * searchTreeMap(TreeMap * tree, void* key) {
     if(tree->root == NULL) return NULL;
     
-    void* (int) keyComp;
+    void* keyComp;
     tree->current = tree->root;
     
     while(tree->current != NULL) {
